# swagger.yml
swagger: "2.0"
info:
  version: 1.0.0
  title: Gym Badges API
...
paths:
...
  /badges/{user_id}:
    get:
      operationId: getBadgesByUserID
      summary: Get badges info by user_id
      tags:
        - Badges
      produces:
        - application/json
      parameters:
        - name: user_id
          in: path
          description: User's id you want to get.
          required: true
          type: string
        - name: auth_user_id
          in: header
          description: Your own user id. For authentication.
          required: true
          type: string
      security:
        - jwt: []
      responses:
        200:
          description: Success Response
          schema:
            $ref: "#/definitions/badges_by_user_response"
        401:
          description: Unauthorized Error
          schema:
            $ref: "#/definitions/generic_response"
            description: Contains the unauthorized error response object
        404:
          description: Not Found Error
          schema:
            $ref: "#/definitions/generic_response"
            description: Contains the not found error response object
        500:
          description: Unexpected Error
          schema:
            $ref: "#/definitions/generic_response"
            description: Contains the operation error response object

    post:
      operationId: AddBadge
      summary: Marks a badge as achieved.
      tags:
        - Badges
      produces:
        - application/json
      parameters:
        - name: user_id
          in: path
          description: Your own user id.
          required: true
          type: string
        - name: auth_user_id
          in: header
          description: Your own user id. For authentication.
          required: true
          type: string
        - name: input
          description: Badge to mark as achieved.
          in: body
          required: true
          schema:
            $ref: "#/definitions/add_delete_badge_request"
      security:
        - jwt: []
      responses:
        200:
          description: Success Response
        401:
          description: Unauthorized Error
          schema:
            $ref: "#/definitions/generic_response"
            description: Contains the unauthorized error response object
        403:
          description: Forbidden error. Returned when user doesn't have the parent badge.
          schema:
            $ref: "#/definitions/generic_response"
            description: Contains the forbidden error response object
        404:
          description: Not Found Error
          schema:
            $ref: "#/definitions/generic_response"
            description: Contains the not found error response object
        500:
          description: Unexpected Error
          schema:
            $ref: "#/definitions/generic_response"
            description: Contains the operation error response object

    delete:
      operationId: DeleteBadge
      summary: Unmarks a badge as achieved.
      tags:
        - Badges
      produces:
        - application/json
      parameters:
        - name: user_id
          in: path
          description: Your own user id.
          required: true
          type: string
        - name: auth_user_id
          in: header
          description: Your own user id. For authentication.
          required: true
          type: string
        - name: input
          description: Badge to unmark as achieved.
          in: body
          required: true
          schema:
            $ref: "#/definitions/add_delete_badge_request"
      security:
        - jwt: []
      responses:
        200:
          description: Success Response
        401:
          description: Unauthorized Error
          schema:
            $ref: "#/definitions/generic_response"
            description: Contains the unauthorized error response object
        403:
          description: Forbidden error. Returned when user doesn't have the parent badge.
          schema:
            $ref: "#/definitions/generic_response"
            description: Contains the forbidden error response object
        404:
          description: Not Found Error
          schema:
            $ref: "#/definitions/generic_response"
            description: Contains the not found error response object
        500:
          description: Unexpected Error
          schema:
            $ref: "#/definitions/generic_response"
            description: Contains the operation error response object

  # -----------------------------------------------------
  # RANKINGS
  # -----------------------------------------------------
  /rankings/global/{user_id}:
    get:
      operationId: getGlobalRanking
      summary: Get global ranking.
      tags:
        - Rankings
      produces:
        - application/json
      parameters:
        - name: user_id
          in: path
          description: User's id you want to get.
          required: true
          type: string
        - name: auth_user_id
          in: header
          description: Your own user id. For authentication.
          required: true
          type: string
        - name: page
          in: query
          description: Page number for pagination (1-based).
          required: true
          type: integer
          format: int32
      security:
        - jwt: []
      responses:
        200:
          description: Success Response
          schema:
            $ref: "#/definitions/get_ranking_response"
        401:
          description: Unauthorized Error
          schema:
            $ref: "#/definitions/generic_response"
            description: Contains the unauthorized error response object
        404:
          description: Not Found Error
          schema:
            $ref: "#/definitions/generic_response"
            description: Contains the not found error response object
        500:
          description: Unexpected Error
          schema:
            $ref: "#/definitions/generic_response"
            description: Contains the operation error response object
  /rankings/friends/{user_id}:
    get:
      operationId: getFriendsRanking
      summary: Get friends ranking.
      tags:
        - Rankings
      produces:
        - application/json
      parameters:
        - name: user_id
          in: path
          description: User's id you want to get.
          required: true
          type: string
        - name: auth_user_id
          in: header
          description: Your own user id. For authentication.
          required: true
          type: string
        - name: page
          in: query
          description: Page number for pagination (1-based).
          required: true
          type: integer
          format: int32
      security:
        - jwt: []
      responses:
        200:
          description: Success Response
          schema:
            $ref: "#/definitions/get_ranking_response"
        401:
          description: Unauthorized Error
          schema:
            $ref: "#/definitions/generic_response"
            description: Contains the unauthorized error response object
        404:
          description: Not Found Error
          schema:
            $ref: "#/definitions/generic_response"
            description: Contains the not found error response object
        500:
          description: Unexpected Error
          schema:
            $ref: "#/definitions/generic_response"
            description: Contains the operation error response object
securityDefinitions:
  jwt:
    type: apiKey
    in: header
    name: token
definitions:
  generic_response:
    type: object
    title: Generic Operation Response
    description: Describes the generic result of a request.
    properties:
      code:
        type: string
      message:
        type: string
...
  badges_by_user_response:
    title: User badge list response
    type: array
    items:
      $ref: "#/definitions/badge"
  badge:
    type: object
    title: User badge
    properties:
      achieved:
        type: boolean
        x-omitempty: false
      description:
        type: string
        x-omitempty: false
      exp:
        type: number
        format: int64
        x-omitempty: false
      id:
        type: number
        format: int32
        x-omitempty: false
      image:
        type: string
        x-omitempty: false
      name:
        type: string
        x-omitempty: false
      children:
        type: array
        items:
          $ref: "#/definitions/badge"
...
  add_delete_badge_request:
    type: object
    title: Badge top be marked as achieved.
    properties:
      badge_id:
        type: number
        format: int32
  get_ranking_response:
    type: object
    title: Ranking response
    properties:
      yourself:
        $ref: "#/definitions/raking_user"
      ranking:
        type: array
        x-omitempty: false
        items:
          $ref: "#/definitions/raking_user"
  raking_user:
    type: object
    title: User info in a ranking.
    properties:
      user_id:
        type: string
        x-omitempty: false
      name:
        type: string
        x-omitempty: false
      image:
        type: string
        format: byte
        x-omitempty: false
      level:
        type: number
        format: int64
        x-omitempty: false
      rank:
        type: number
        format: int64
        x-omitempty: false
      streak:
        type: number
        format: int32
        x-omitempty: false